Ability to delete by tag
Yes, I know it's a race condition in the big version of the repo, but
it shouldn't be an issue here. The ability to not delete a particular
tag without deleting all of them that point to a manifest is a big
sticking point for us.

It still works like the reference version by falling back to the
SHA of the manifest if you're so inclined to use it.

Fixes #20